+++ SightKeeper.Services.Windows/WindowsMouseMover.cs
@@ -6,18 +6,18 @@ namespace SightKeeper.Services.Windows;
 
 public sealed class WindowsMouseMover : MouseMover
 {
-    public IObservable<(short xDelta, short yDelta)> Moved => _moved.AsObservable();
+    public IObservable<(int xDelta, int yDelta)> Moved => _moved.AsObservable();
 
-    public void Move(short xDelta, short yDelta)
+    public void Move(int xDelta, int yDelta)
     {
         User32.MouseMove(xDelta, yDelta);
         _moved.OnNext((xDelta, yDelta));
     }
 
-    public void SetPosition(short x, short y)
+    public void SetPosition(int x, int y)
     {
         Cursor.Position = new Point(x, y);
     }
 
-    private readonly Subject<(short, short)> _moved = new();
+    private readonly Subject<(int, int)> _moved = new();
 }
